Ignore special type properties of native functions
We no longer take .prototype, ... from native functions,
because that seems to break sometimes.

"Type Error: Illegal Invocation"

No tests added for this case, since the only known example
for this is the `URL` object is Browser only.
This did not occur with other native objects, like Array or
Object.