Add a stand-alone dockerized node service for object-hash
I realized that this tool would be helpful to call from tools such as JMeter and Postman. Of course, since JMeter is written in Java, the container for that would be built primarily for Java. Having a service just for object-hash seemed to make sense.

This is not the best node code, but it is a good start.