Update readme to detail passthrough mode. (#107)
This resolves issue #106 .

I was looking for an algorithm option which would display the unhashed value, read the documentation, and assumed there wasn't one, without a more thorough look at the code.

Once I inspected the code to make a pull request, I saw the feature already existed. The simple solution is to describe the feature in the appropriate place in the documentation of `hash()`.